"HOT FIX : set global variable for cli execution in separate function
set global variable for cli execution in separate function
so that imports in migration_cycle can be at the top.
Once the global_vars file is loaded it cannot be changed that's why
import was added later.
to fix that simply set the global in migration_manager before
executing host_migration